+++ apps/portal/jest.config.js
@@ -5,6 +5,10 @@ module.exports = {
     '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nrwl/react/plugins/jest',
     '^.+\\.[tj]sx?$': 'babel-jest'
   },
+  setupFilesAfterEnv: [
+    '@testing-library/jest-dom/extend-expect',
+    './setupTests.js'
+  ],
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../../coverage/apps/portal'
 }