+++ apps/portal/src/views/Dashboard/index.spec.tsx
@@ -1,12 +1,10 @@
 import { MemoryRouter } from 'react-router-dom'
 import { render } from '@testing-library/react'
 
-import '../../helpers/__mocks__/matchMedia'
-
 import Dashboard from '.'
 
 // eslint-disable-next-line react/jsx-no-useless-fragment
-jest.mock('../../components/MapboxGL', () => () => <></>)
+jest.mock('../../../../../libs/shared/ui/src/lib/MapboxGL', () => () => <></>)
 
 describe('Dashboard', () => {
   it('should render successfully', () => {