+++ apps/portal/src/views/Location/index.spec.tsx
@@ -1,7 +1,9 @@
-import { render } from '@testing-library/react'
+import { render } from '@phc/shared-utils'
 
 import Location from '.'
 
+// eslint-disable-next-line react/jsx-no-useless-fragment
+jest.mock('../../../../../libs/shared/ui/src/lib/MapboxGL', () => () => <></>)
 describe('Location', () => {
   it('should render successfully', () => {
     const { baseElement } = render(<Location />)