+++ apps/portal/src/views/Location/index.tsx
@@ -1,4 +1,4 @@
-import { Link, Route } from 'react-router-dom'
+// import { Link, Route } from 'react-router-dom'
 import styled from 'styled-components'
 
 /* eslint-disable-next-line */
@@ -12,13 +12,6 @@ export function Location(props: LocationProps) {
   return (
     <StyledLocation>
       <h1>Welcome to Location!</h1>
-
-      <ul>
-        <li>
-          <Link to="/">Location root</Link>
-        </li>
-      </ul>
-      <Route path="/" element={<div>This is the Location root route.</div>} />
     </StyledLocation>
   )
 }