+++ apps/portal/src/views/Settings/index.spec.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react'
 import Settings from '.'
 
 // eslint-disable-next-line react/jsx-no-useless-fragment
-jest.mock('../../components/MapboxGL', () => () => <></>)
+jest.mock('../../../../../libs/shared/ui/src/lib/MapboxGL', () => () => <></>)
 
 describe('Settings', () => {
   it('should render successfully', () => {