+++ apps/portal/src/views/Users/index.spec.tsx
@@ -1,11 +1,11 @@
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { MemoryRouter } from 'react-router-dom'
-import { render } from '@testing-library/react'
+import { render } from '@phc/shared-utils'
 
 import Users from '.'
 
 // eslint-disable-next-line react/jsx-no-useless-fragment
-jest.mock('../../components/MapboxGL', () => () => <></>)
+jest.mock('../../../../../libs/shared/ui/src/lib/MapboxGL', () => () => <></>)
 
 const queryClient = new QueryClient()