+++ apps/shared-ui-e2e/src/integration/index/index.spec.ts
@@ -1,7 +1,0 @@
-describe('shared-ui: UserProfile component', () => {
-  beforeEach(() => cy.visit('/iframe.html?id=userprofile--primary'))
-
-  it('should render the component', () => {
-    cy.get('h1').should('contain', 'Welcome to UserProfile!')
-  })
-})