+++ libs/shared/hooks/src/index.ts
@@ -1,1 +1,1 @@
-// export * from './lib/use-request'
+// export * from './lib/use-autocomplete-input'